Juancho Sincro Clase

+++ Admin/Admin_Modificar_img.php
@@ -2,18 +2,9 @@
 session_start();
 
 	require '../Inclu/Admin_Inclu_01b.php';
-
-		require '../Conections/conection.php';
-
-	$db = mysqli_connect($db_host,$db_user,$db_pass,$db_name);
-	if (!$db){ die ("Es imposible conectar con la bbdd ".$db_name."</br>".mysqli_connect_error());
-				}
-
-	$sqld =  "SELECT * FROM `admin` WHERE `Email` = '$_POST[Email]' OR `Usuario` = '$_POST[Usuario]'";
- 	
-	$qd = mysqli_query($db, $sqld);
-	
-	$rowd = mysqli_fetch_assoc($qd);
+	require '../Conections/conection.php';
+	require '../Conections/conexion_bbdd.php';
+	require 'Admin_select_rowd.php';
 
 ///////////////////////////////////////////////////////////////////////////////////////
 
@@ -22,10 +13,11 @@ if ($_SESSION['Nivel'] == 'XBPadmin'){
  					print("Hello ".$_SESSION['Nombre']." ".$_SESSION['Apellidos'].".");
 							master_index();
 					
-							if ($_POST['oculto2']){
+							if (isset($_POST['oculto2'])){
 								show_form();
 								}
-							elseif($_POST['imagenmodif']){
+							elseif(isset($_POST['imagenmodif'])
+							){
 								
 									if($form_errors = validate_form()){
 										show_form($form_errors);
@@ -35,22 +27,7 @@ if ($_SESSION['Nivel'] == 'XBPadmin'){
 								} else {
 											show_form();
 									}
-				} else { 
-					
-						print("<table align='center' style=\"margin-top:200px;margin-bottom:200px\">
-									<tr align='center'>
-										<td>
-											<font color='red'>
-												<b>
-													ACCESO RESTRINGIDO.
-												</br></br>
-													CONSULTE SUS PERMISOS ADMINISTRATIVOS.
-											</font>
-										</td>
-									</tr>
-								</table>");
-								
-							}
+				} else { require 'Admin_denegado.php'; }
 
 //////////////////////////////////////////////////////////////////////////////////////////////
 
@@ -142,7 +119,9 @@ function process_form(){
 	$fget_old = trim($fget_old);
 	fclose($fw_old);
 	$_SESSION['img_old2'] = $fget_old;
-	unlink ("img/Old".$_SESSION['img_old2']);
+
+	if(file_exists("img/Old".$_SESSION['img_old2'])){ unlink ("img/Old".$_SESSION['img_old2']); } else { }
+	
 	unset ($_SESSION['img_old2']);
 	
 	if(file_exists('img/old_img.txt')){ unlink("img/old_img.txt");}
@@ -207,7 +186,7 @@ function process_form(){
 
 //////////////////////////////////////////////////////////////////////////////////////////////
 
-function show_form($errors=''){
+function show_form($errors=[]){
 	
 	global $db; 	
 	
@@ -287,14 +266,18 @@ function show_form($errors=''){
 
 /////////////////////////////////////////////////////////////////////////////////////////////////
 	
-	function master_index(){
-		
-				require '../Inclu/Master_Index_Admin.php';
-				
-				} 
+function master_index(){
+
+	global $RutaDir;
+	$RutaDir = "Admin";
+	require '../Inclu/Master_Index.php';
+			
+	} 
 
 /////////////////////////////////////////////////////////////////////////////////////////////////
 	
-	require '../Inclu/Admin_Inclu_02.php';
+	require '../Inclu/Inclu_Footer.php';
+
+					 /* Creado por Juan Manuel Barrós Pazos 2008/2022 */
 		
 ?>
